chore(code): flatten and rename fw provider package
