+++ fwprovider/datasource_version.go
@@ -4,7 +4,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package provider
+package fwprovider
 
 import (
 	"context"
