+++ fwprovider/structure/attribute.go
