+++ fwprovider/tests/datasource_version_test.go
@@ -4,7 +4,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package test
+package tests
 
 import (
 	"context"
