+++ fwprovider/tests/resource_linux_bridge_test.go
@@ -4,21 +4,19 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package network
+package tests
 
 import (
 	"context"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-
-	"github.com/bpg/terraform-provider-proxmox/internal/test"
 )
 
 func TestLinuxBridgeResource(t *testing.T) {
 	t.Parallel()
 
-	accProviders := test.AccMuxProviders(context.Background(), t)
+	accProviders := AccMuxProviders(context.Background(), t)
 
 	resourceName := "proxmox_virtual_environment_network_linux_bridge.test"
 
@@ -27,7 +25,7 @@ func TestLinuxBridgeResource(t *testing.T) {
 		Steps: []resource.TestStep{
 			// Create and Read testing
 			{
-				Config: test.ProviderConfig + `
+				Config: ProviderConfig + `
 resource "proxmox_virtual_environment_network_linux_bridge" "test" {
 	node_name = "pve"
 	name = "vmbr99"
@@ -56,7 +54,7 @@ resource "proxmox_virtual_environment_network_linux_bridge" "test" {
 			},
 			// Update testing
 			{
-				Config: test.ProviderConfig + `
+				Config: ProviderConfig + `
 resource "proxmox_virtual_environment_network_linux_bridge" "test" {	
 	node_name = "pve"
 	name = "vmbr99"
@@ -81,7 +79,7 @@ resource "proxmox_virtual_environment_network_linux_bridge" "test" {
 			},
 			// Create with other default overrides
 			{
-				Config: test.ProviderConfig + `
+				Config: ProviderConfig + `
 resource "proxmox_virtual_environment_network_linux_bridge" "test" {
 	node_name = "pve"
 	name = "vmbr98"
