+++ fwprovider/tests/resource_linux_vlan_test.go
@@ -4,21 +4,19 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package network
+package tests
 
 import (
 	"context"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-
-	"github.com/bpg/terraform-provider-proxmox/internal/test"
 )
 
 func TestLinuxVLANResource(t *testing.T) {
 	t.Parallel()
 
-	accProviders := test.AccMuxProviders(context.Background(), t)
+	accProviders := AccMuxProviders(context.Background(), t)
 
 	resourceName := "proxmox_virtual_environment_network_linux_vlan.test"
 
@@ -27,7 +25,7 @@ func TestLinuxVLANResource(t *testing.T) {
 		Steps: []resource.TestStep{
 			// Create and Read testing
 			{
-				Config: test.ProviderConfig + `
+				Config: ProviderConfig + `
 resource "proxmox_virtual_environment_network_linux_vlan" "test" {
 	node_name = "pve"
 	name = "eno0.33"
@@ -51,7 +49,7 @@ resource "proxmox_virtual_environment_network_linux_vlan" "test" {
 			},
 			// Update testing
 			{
-				Config: test.ProviderConfig + `
+				Config: ProviderConfig + `
 resource "proxmox_virtual_environment_network_linux_vlan" "test" {
 	node_name = "pve"
 	name = "eno0.33"
