+++ fwprovider/validators/ha_resource_validator.go
