--- internal/internal.go
@@ -1,8 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- */
-
-// Package internal contains implementation of the provider created with Terraform Plugin Framework.
-package internal
